Changes to rendering to accomodate iOS. Factored in devicePerPixelRatio and improved vsync on iOS.

+++ shaders/default_vert.glsl
@@ -10,7 +10,7 @@ precision mediump float;
 // Input from Engine
 attribute highp     vec4 vertex;                    // Input
 attribute highp     vec4 texture_coordinates;       // Input
-uniform   mediump   mat4 u_matrix;                  // Input
+uniform   highp     mat4 u_matrix;                  // Input
 
 // Output to Fragment Shader
 varying   highp     vec4 coordinates;               // Output
